% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black.white.R
\name{black.white}
\alias{black.white}
\title{black.white}
\usage{
black.white(n)
}
\arguments{
\item{n}{Number of colors in palette.}
}
\value{
A	Character vector of colors
}
\description{
Create a vector of n continuous colors from black to white
}
\examples{
plot(1:1000,1:1000, col=black.white(1000), pch=19)
}
\seealso{
\link[grDevices]{colorRampPalette}
}

