% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jet.colors.R
\name{jet.colors}
\alias{jet.colors}
\title{jet.colors}
\usage{
jet.colors(n)
}
\arguments{
\item{n}{Number of colors in palette.}
}
\value{
A	Character vector of colors
}
\description{
Create a vector of n continuous colors from blue to red
}
\examples{
plot(1:1000,1:1000, col=jet.colors(1000), pch=19)
}
\seealso{
\link[grDevices]{colorRampPalette}
}

