% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layerAvg.R
\name{layerAvg}
\alias{layerAvg}
\title{layerAvg}
\usage{
layerAvg(x, y, x1, x2)
}
\arguments{
\item{x}{Sorted vector of the independent variable}

\item{y}{A vector of the dependent variable}

\item{x1}{Lower bound of x}

\item{x2}{Upper bound of x}
}
\value{
The integrated mean of y
}
\description{
Calculates the integrated mean of y within the defined bounds of x
}
\examples{
x <- c(1:10)
y <- x^2
layerAvg(x, y, 2, 8)
}

