% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mNine.R
\docType{data}
\name{mNine}
\alias{mNine}
\title{mNine}
\format{\code{mNine} is a \link[base]{list} twenty MATLAB files imported as lists. Each MATLAB file list contains data collected with a SonTek m9 ADCP, which was exported from RiverSurveyor Live as a MATLAB file. Although data were collected with an m9, they have been edited to represent fictitious values and locations.}
\usage{
\code{data(mNine)}
}
\description{
Dataset of SonTek ADCP transects exported from RiverSurveyor Live as MATLAB files
}
\details{
\itemize{
\item \code{t1l}	ADCP data from transect one replicate starting at left bank
\item \code{t1r}	ADCP data from transect one replicate starting at right bank
\item \code{t2l}	ADCP data from transect two replicate starting at left bank
\item \code{t2r}	ADCP data from transect two replicate starting at right bank
\item \code{t3l}	ADCP data from transect three replicate starting at left bank
\item \code{t3r}	ADCP data from transect three replicate starting at right bank
\item \code{t4l}	ADCP data from transect four replicate starting at left bank
\item \code{t4r}	ADCP data from transect four replicate starting at right bank
\item \code{t5l}	ADCP data from transect five replicate starting at left bank
\item \code{t5r}	ADCP data from transect five replicate starting at right bank
\item \code{t6l}	ADCP data from transect six replicate starting at left bank
\item \code{t6r}	ADCP data from transect six replicate starting at right bank
\item \code{t7l}	ADCP data from transect seven replicate starting at left bank
\item \code{t7r}	ADCP data from transect seven replicate starting at right bank
\item \code{t8l}	ADCP data from transect eight replicate starting at left bank
\item \code{t8r}	ADCP data from transect eight replicate starting at right bank
\item \code{t9l}	ADCP data from transect nine replicate starting at left bank
\item \code{t9r}	ADCP data from transect nine replicate starting at right bank
\item \code{t10l}	ADCP data from transect ten replicate starting at left bank
\item \code{t10r}	ADCP data from transect ten replicate starting at right bank
}
}
\keyword{dataset}

