% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho.proj.R
\name{ortho.proj}
\alias{ortho.proj}
\title{ortho.proj}
\usage{
ortho.proj(x, y, m, b)
}
\arguments{
\item{x}{A vector of X geographic coordinates}

\item{y}{A vector of Y geographic coordinates}

\item{m}{Slope parameter from linear regression of y~x}

\item{b}{Intercept parameter from linear regression of y~x}
}
\value{
A \link[base]{data.frame} containing projected x and y coordinates (xProj and yProj). The number of rows equals the number of observations in \code{x}.
}
\description{
Calculates the x and y coordinates along an orthogonal plane, given the linear regression coefficients of y~x
}
\examples{
data(vels)
vels.t1 <- vels[transectName=="t1",,]
plot(UTM_Y~UTM_X, vels.t1)
linmod <- lm(vels.t1$UTM_Y~vels.t1$UTM_X)
projected <- ortho.proj(vels.t1$UTM_X, vels.t1$UTM_Y, linmod$coefficients[2], linmod$coefficients[1])
plot(projected)
}

