% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.planform.R
\name{plot.planform}
\alias{plot.planform}
\title{plot.planform}
\usage{
plot.planform(data, x, interpolate, coord = "projected.utm",
  point.spacing = 1, point.color = jet.colors, arrows = TRUE,
  arrow_scale = 1, arrowwd = 1, arrow.key, keyCol = "black",
  textCol = "black", xUnits = NULL, cellWidth, cellHeight,
  interpolate.color = jet.colors, xlab = list(label = "X Coordinates",
  fontsize = 18), ylab = list(label = "Y Coordinates", fontsize = 18),
  ylab.right = list(label = as.character(interpolate), fontsize = 18),
  colorkey = list(labels = list(cex = 1.25)), scales = list(cex = c(1.25,
  1.25)), key.cex = 1.25, xlim, ylim, aspect = "iso",
  par.settings = list(layout.widths = list(axis.key.padding = 0, ylab.right =
  2)), ..., contour = TRUE, cuts = 7, labels = TRUE, contourwd = 1)
}
\arguments{
\item{data}{An object of class "adcp.planform"}

\item{x}{Parameter to be plotted as points using \code{xyplot}}

\item{interpolate}{Optional. Parameter to plot using \code{levelplot}, if provided in addition to \code{x} the interpolated parameter is overlaid by \code{x}.}

\item{coord}{Spatial coordinates used to define x and y-axes, can be orthogonally "projected.utm" (default), "utm", orthogonally "projected.latlong", or "latlong".}

\item{point.spacing}{Distance between points, default is 1.}

\item{point.color}{Color of points displayed, maybe numeric, character string, or function. Default is \code{jet.colors}.}

\item{arrows}{If TRUE (default) and x is a velocity vector, arrows indicating flow direction are plotted}

\item{arrow_scale}{Multiplication factor for sizing secondary flow arrows, default is 1}

\item{arrowwd}{Line width of arrows}

\item{arrow.key}{Numeric vector of length 2 defining the X and y coordinates (c(x,y)) of where scaling key for points or arrows (if \code{arrows} = TRUE) should be written. If omitted, a key is placed in the lower left.}

\item{keyCol}{Color of points or arrows (if \code{arrows} = TRUE) in arrow.key, default is "black".}

\item{textCol}{Colors of text in arrow.key, default is "black".}

\item{xUnits}{Character string. If provided, the units of \code{x} (e.g., "m" or "m/s") to be displayed in arrow.key.}

\item{cellWidth}{Width of cell size used to grid interpolate. If missing, defaults to 1/100 of the range of \code{xlim}.}

\item{cellHeight}{Height of cell size used to grid interpolate. If missing, defaults to 1/100 of the range of \code{ylim}.}

\item{interpolate.color}{Color function used to display \code{interpolate}, see \link[lattice]{levelplot} for details. Default is \code{jet.colots}.}

\item{xlab}{X-axis label, see \link[lattice]{levelplot} for details. Default is "X Coordinates" in size 18 font.}

\item{ylab}{Y-axis label, see \code{levelplot} for details. Default is "Y Coordinates" in size 18 font.}

\item{ylab.right}{\code{colorkey} label, see \link[lattice]{xyplot} for details. Default is as.character(interpolate) in size 18 font.}

\item{colorkey}{A list of arguments for the color key drawn alongside the plot, see \link[lattice]{levelplot} for details. Default writes the tick labels with cex 1.5.}

\item{scales}{A list of arguments determining how the axes are drawn, see \link[lattice]{xyplot} for details. Default is to draw tick labels with cex 1.5.}

\item{key.cex}{Font size of text in arrow.key, default is 1.25}

\item{xlim}{Numeric vector of length 2 defining the range of the x-axis.}

\item{ylim}{Numeric vector of length 2 defining the range of the y-axis.}

\item{aspect}{Controls the aspect ratio of the panels, see \link[lattice]{xyplot} for details. Default is "iso".}

\item{par.settings}{A list of arguments for fine-tuned control of display, see \link[lattice]{xyplot} and \link[lattice]{trellis.par.set} for details. Default is not to pad the color key, offset the color key label by 2.}

\item{...}{Additional arguments to be passed to \link[lattice]{xyplot} and \link[lattice]{levelplot}.}

\item{contour}{Logical, if TRUE (default) and \code{interpolate} is defined, contour lines are drawn}

\item{cuts}{Number of regions contour lines denote}

\item{labels}{Logical, if TRUE (default) contour lines are labeled}

\item{contourwd}{Line width of contours}
}
\value{
A spatially referenced plot of \code{x} and \code{interpolate} (if provided).
}
\description{
Plots spatial data from an object of class "adcp.planform"
}
\examples{
data(mNine)
#mNine is a list of MATLAB files
names(mNine)
#Drop the last letter (l or r) from the MATLAB file names
tNames <- substr(names(mNine), 0, nchar(names(mNine))-1)
adcp.planform <- process.planform(mNine, tNames, Dc = 0.0375, n = 2, xWindow = 21)
#Plot depth averaged speed with arrows displaying flow heading
plot.planform(adcp.planform, DepthAveragedSpeed, arrow.key = c(325540,4679130), xUnits = "m/s", arrow_scale = 10)
#Plot depth averaged speed as points
plot.planform(adcp.planform, DepthAveragedSpeed, arrow.key = c(325540,4679130), xUnits = "m/s", arrows = FALSE)
#Plot depth averaged speed overlaying depth
plot.planform(adcp.planform, DepthAveragedSpeed, interpolate = depth, arrow.key = c(325540,4679130), xUnits = "m/s", point.spacing = 10, point.color = "black", ylab.right = list(label = "Depth (m)", fontsize = 18), arrow_scale=10)
}
\seealso{
\link{process.planform},\link[lattice]{levelplot}, and \link[lattice]{xyplot}.
}

