% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.xSec.default.R
\name{plot.xSec.default}
\alias{plot.xSec.default}
\title{plot.xSec.default}
\usage{
plot.xSec.default(data, x, transect, cellWidth, cellHeight, arrows = TRUE,
  arrow.x.spacing, arrow.y.spacing, arrow_scale = 1, arrowCol = "black",
  arrowwd = 1, lwd = 1, arrow.key, keyCol = "white", lineCol = "white",
  col.regions = jet.colors, scale.by = "individual", xlab = list(label =
  "Distance (m)", fontsize = 18), ylab = list(label = "Depth (m)", fontsize =
  18), ylab.right = list(label = "Velocity (m/s)", fontsize = 18),
  header = transect, headersize = 18, colorkey = list(labels = list(cex =
  1.5)), scales = list(cex = c(1.5, 1.5)), key.cex = 1.25, xlim, ylim,
  aspect = "iso", par.settings = list(layout.widths = list(axis.key.padding
  = 0, ylab.right = 2), panel.background = list(col = "black")), ...,
  contour = TRUE, cuts = 7, labels = TRUE)
}
\arguments{
\item{data}{An object of class "adcp.secondary"}

\item{x}{Parameter to be plotted}

\item{transect}{Transect to be plotted}

\item{cellWidth}{Width of cell size used to grid x. If missing, defaults to mean cell width from dataset}

\item{cellHeight}{Height of cell size used to grid x. If missing, defaults to mean cell height from dataset}

\item{arrows}{If TRUE (default) and x is a primary velocity vector, arrows indicating secondary flow velocity are plotted}

\item{arrow.x.spacing}{Horizontal spacing between secondary flow arrows. Units equal to units of x-axis (i.e., distance from start of transect).}

\item{arrow.y.spacing}{Vertical spacing secondary flow arrows. Units equal to units of y-axis (i.e., depth).}

\item{arrow_scale}{Multiplication factor for sizing secondary flow arrows, default is 1}

\item{arrowCol}{Color of secondary flow arrows, default is "black"}

\item{arrowwd}{Line width of arrows}

\item{lwd}{Line width of surface and bottom outlines}

\item{arrow.key}{Numeric vector of length 2 defining the X and y coordinates (c(x,y)) of where scaling key for secondary flow arrows should be written. If omitted, a key is placed in the lower left.}

\item{keyCol}{Color of text and arrows in arrow.key, default is "white"}

\item{lineCol}{Line color of surface and bottom outlines, default is white}

\item{col.regions}{Color function used to display x, see \link[lattice]{levelplot} for details. Default is \link{jet.colors}.}

\item{scale.by}{Character string. If "individual" (default) variables in plot are scaled to their range in the cross-section plotted. If "all" variables in plot are scaled to their range in the dataset provided. See 'details' for additional information}

\item{xlab}{X-axis label, see \link[lattice]{levelplot} for details. Default is "Distance (m)" in size 18 font.}

\item{ylab}{Y-axis label, see \link[lattice]{levelplot} for details. Default is "Depth (m)" in size 18 font.}

\item{ylab.right}{\code{colorkey} label, see \link[lattice]{xyplot} for details. Default is "Velocity m/s" in size 18 font.}

\item{header}{Title of plot, default is transect}

\item{headersize}{Font size of header, default is 18}

\item{colorkey}{A list of arguments for the color key drawn alongside the plot, see \link[lattice]{levelplot} for details. Default writes the tick labels with cex 1.5.}

\item{scales}{A list of arguments determining how the axes are drawn, see \link[lattice]{xyplot} for details. Default is to draw tick labels with cex 1.5.}

\item{key.cex}{Font size of text in arrow.key, default is 1.25}

\item{xlim}{Numeric vector of length 2 defining the range of the x-axis.}

\item{ylim}{Numeric vector of length 2 defining the range of the y-axis. See 'details' for additional information.}

\item{aspect}{Controls the aspect ratio of the panels, see \link[lattice]{xyplot} for details. Default is "iso".}

\item{par.settings}{A list of arguments for fine-tuned control of display, see \link[lattice]{xyplot} and \link[lattice]{trellis.par.set} for details. Default is not to pad the color key, offset the color key label by 2, and draw a black background.}

\item{...}{Additional arguments to be passed to \link[lattice]{levelplot}.}

\item{contour}{Logical, if TRUE (default), contour lines are drawn}

\item{cuts}{Number of regions contour lines denote}

\item{labels}{Logical, if TRUE (default) contour lines are labeled}
}
\value{
A levelplot plot of the cross-section
}
\description{
Plots cross-section velocities given an object of class "adcp.secondary"
}
\details{
When producing multiple plots of cross-sections, setting scale.by to "all" eases comparison among plots by using the same scale for all plots.  Setting scale.by to "individual" exaggerates differences within individual cross-sections. 
If scale.by equals "individual" default for \code{ylim} is the max depth of the cross-section to -0.5. If scale.by equals "all" default for \code{ylim} is the max depth of the dataset to -0.5.
Use transect argument to subset data by transect, may produce error if data is subset prior to being supplied to function.
}
\examples{
data(mNine)
#mNine is a list of MATLAB files
names(mNine)
#Drop the last letter (l or r) from the MATLAB file names
tNames <- substr(names(mNine), 0, nchar(names(mNine))-1)
adcp.secondary <- process.secondary(mNine, tNames, rotation = c("xSec", "rozovskii", "zeroSecQ"), xWindow = 21, yWindow = 5)
plot.xSec(adcp.secondary, vp.zsq, transect = "t1")
#Use the same scale to display primary and secondary velocities of two cross-sections
plot.xSec(adcp.secondary, vp.zsq, transect = "t1", scale.by = "all")
}
\seealso{
\link{plot.xSec}, \link[lattice]{levelplot}, \link{process.secondary}, and \link[lattice]{xyplot}.
}

