% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.xSec.R
\name{plot.xSec}
\alias{plot.xSec}
\title{plot.xSec}
\usage{
plot.xSec(data, x, transects, ...)
}
\arguments{
\item{data}{An object of class "adcp.secondary"}

\item{x}{Parameter to be plotted}

\item{transects}{A list of transects to be plotted, if missing all transects are plotted}

\item{...}{Arguments to be passed to \link{plot.xSec.default}.}
}
\value{
A list of levelplot plots. Transect names contained within the dataset are used to name each element (i.e., plot) within the list.
}
\description{
Plots cross-section velocities given an object of class "adcp.secondary"
}
\details{
Use transects argument to subset data by transect, may produce error if data is subset prior to being supplied to function.
}
\examples{
data(mNine)
#mNine is a list of MATLAB files
names(mNine)
#Drop the last letter (l or r) from the MATLAB file names
tNames <- substr(names(mNine), 0, nchar(names(mNine))-1)
adcp.secondary <- process.secondary(mNine, tNames, rotation = c("xSec", "rozovskii", "zeroSecQ"), xWindow = 21, yWindow = 5)
secondaryQ.zsq <- plot.xSec(adcp.secondary, vp.zsq)
#Display first cross-section
secondaryQ.zsq[[1]]
#Alternatively
secondaryQ.zsq$t1
}
\seealso{
\link{plot.xSec.default}, \link[lattice]{levelplot}, \link{process.secondary}.
}

