% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project.transect.R
\name{project.transect}
\alias{project.transect}
\title{project.transect}
\usage{
project.transect(x, y, transectNames)
}
\arguments{
\item{x}{A vector of X geographic coordinates}

\item{transectNames}{A vector identifying the transect corresponding to each x and y coordinate. Individual replicates should not be identified, see 'details' for additional information.}

\item{z}{Vertical elevation  coordinates}
}
\value{
A \link[data.table]{data.table} of sample id, transect names, x coordinates fitted to an orthogonal plane (xProj), and y coordinates fitted to an orthogonal plane (yProj).
}
\description{
Orthogonally projects x and y coordinates of replicated ADCP transects to a mean transect line.
}
\details{
\code{transectNames} is used to group replicate transects together during processing, if each replicate has a unique name all replicates will be handled as unique transects and projected coordinates will not represent the mean transect line. If \code{transectNames} is missing, all coordinates are assumed to correspond to a single transect.
}
\examples{
data(vels)
plot(UTM_Y~UTM_X, vels)
mNineProject <- project.transect(vels$UTM_X, vels$UTM_Y, vels$transectName)
plot(yProj~xProj, mNineProject)
}
\seealso{
\link{ortho.proj}
}

