% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozovskii.R
\name{rozovskii}
\alias{rozovskii}
\title{rozovskii}
\usage{
rozovskii(ve, vn, vu, meanVe, meanVn)
}
\arguments{
\item{ve}{Vector of east velocities within cells}

\item{vn}{Vector of north velocities within cells. Must be the same length as \code{ve}.}

\item{vu}{Vector of vertical velocities within cells. Must be the same length as \code{ve}.}

\item{meanVe}{Vector of depth averaged east velocities of the ensemble a cell belongs to. Must be the same length as \code{ve}.}

\item{meanVn}{Vector of depth averaged north velocities of the ensemble a cell belongs to. Must be the same length as \code{ve}.}
}
\value{
Returns a \link[base]{data.frame} with heading of primary velocity (VpHeading.roz), compass heading of secondary velocity (vsCompassHeading.roz), arithmetic heading of secondary velocity (vsArithmeticHeading.roz), difference between cell heading and primary velocity heading (Theta.roz), primary velocity (vp.roz), secondary velocity (vs.roz), cross-stream component of primary velocity (vpy.roz), down-stream component of primary velocity (vpx.roz), cross-stream component secondary velocity (vsy.roz), and down-stream component of secondary velocity (vsx.roz)
}
\description{
Uses the Rozovskii method of rotation to calculate secondary velocities within a cross-section
}
\examples{
data(cellVels)
velSub <- cellVels[transectName == "t1",,]
roz <- rozovskii(velSub$Vel.E, velSub$Vel.N, velSub$Vel.up, velSub$Mean.Vel.E, velSub$Mean.Vel.N)
}
\references{
Rhoads, B.L. and S.T. Kenworthy. 1998. Time-averaged flow structure in the central region of a stream confluence. Earth Surface Processes and Landforms 23:171-191.
}

