% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatialAverage.R
\name{spatialAverage}
\alias{spatialAverage}
\title{spatialAverage}
\usage{
spatialAverage(data, xWindow, yWindow, spatialCoords, groups,
  na.ignore = FALSE)
}
\arguments{
\item{data}{Dataset containing data to be averaged and x, y coordinates}

\item{xWindow}{Number of neighbors to include in the x orientation of the moving window, including the observation the window centers on.}

\item{yWindow}{Optional. Number of neighbors to include in the y orientation of the moving window, including the observation the window centers on.}

\item{spatialCoords}{A vector of column name(s) containing spatial coordinates within the dataset. The x coordinates must be designated first, followed by the y coordinates. If \code{yWindow} is not provided, the y coordinates can be omitted, otherwise they will be ignored. If \code{spatialCoords} is omitted, a moving window will be used, but data will not be spatially referenced prior to calculations.}

\item{groups}{Optional. Names of column(s) containing information used to subset the data into separate groups for processing. Useful if dataset contains multiple discrete study areas.}

\item{na.ignore}{Logical. If FALSE (default) NA values are assumed to be the average of non-NA values within the moving window, if TRUE NA values remain NA.}
}
\value{
A \link[data.table]{data.table} of spatially averaged values
}
\description{
Averages spatially referenced data using a moving window
}
\examples{
data(vels)
#Average velocities along ADCP transects
velSpatialMean <- spatialAverage(vels, 21, spatialCoords = "tDist", groups = "transectName")
#Average velocities along ADCP transects and ensembles
data(cellVels)
velSpatialMean <- spatialAverage(cellVels, 21, 5, spatialCoords = c("tDist","cellDepth"), groups = "transectName")
}
\seealso{
\link[caTools]{runmean}
}

