% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSD.R
\name{spatialSD}
\alias{spatialSD}
\title{spatialSD}
\usage{
spatialSD(data, xWindow, yWindow, spatialCoords, groups, na.ignore = FALSE)
}
\arguments{
\item{data}{Dataset containing data used to calculate standard deviation and x, y coordinates.}

\item{xWindow}{Number of neighbors to include in the x orientation of the moving window, including the observation the window centers on.}

\item{yWindow}{Optional. Number of neighbors to include in the y orientation of the moving window, including the observation the window centers on.}

\item{spatialCoords}{A vector of column name(s) containing spatial coordinates within the dataset. The x coordinates must be designated first, followed by the y coordinates. If \code{yWindow} is not provided, the y coordinates can be omitted, otherwise they will be ignored. If \code{spatialCoords} is omitted, a moving window will be used, but data will not be spatially referenced prior to calculations.}

\item{groups}{Optional. Names of column(s) containing information used to subset the data into seperate groups for processing. Useful if dataset contains multiple discrete study areas.}

\item{na.ignore}{Logical. If FALSE (default) NA values are assumed to be the standard deviation of non-NA values within the moving window, if TRUE NA values remain NA.}
}
\value{
A \link[data.table]{data.table} of spatially calculated standard deviation values
}
\description{
Computes the standard deviation of spatially referenced data using a moving window
}
\examples{
data(vels)
#Standard deviation of velocities along ADCP transects
velSpatialSD <- spatialSD(vels, 21, spatialCoords = "tDist", groups = "transectName")
#Standard devation of velocities along ADCP transects and ensambles
data(cellVels)
velSpatialSD <- spatialSD(cellVels, 21, 5, spatialCoords = c("tDist","cellDepth"), groups = "transectName")
}
\seealso{
\link[caTools]{runsd}
}

