% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transectHeading.R
\name{transectHeading}
\alias{transectHeading}
\title{transectHeading}
\usage{
transectHeading(x, y, velE, velN, depth, flowHeading)
}
\arguments{
\item{x}{A vector of x coordinates.}

\item{y}{A vector of y coordinates. Must be the same length as \code{x}.}

\item{velE}{A vector of velocities to the east. Must be the same length as \code{x}.}

\item{velN}{A vector of velocities to the north. Must be the same length as \code{x}.}

\item{depth}{A vector of water depths. Must be the same length as \code{x}.}

\item{flowHeading}{Optional. Mean direction of flow over the cross-section. If provided velE, velN, and depth may be omitted.}
}
\value{
Returns the compass heading of a transect
}
\description{
Determines the compass heading of a cross-section transect, assuming transect starts at river right
}
\examples{
data(vels)
velSub <- vels[transectName == "t1",,]
transectHeading(velSub$UTM_X_Proj, velSub$UTM_Y_Proj, velSub$Mean.Vel.E, velSub$Mean.Vel.N, velSub$depth)
#If flow heading is known
transectHeading(velSub$UTM_X_Proj, velSub$UTM_Y_Proj, flowHeading = 149)
}

