% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/white.black.R
\name{white.black}
\alias{white.black}
\title{white.black}
\usage{
white.black(n)
}
\arguments{
\item{n}{Number of colors in palette.}
}
\value{
A	Character vector of colors
}
\description{
Create a vector of n continuous colors from white to black
}
\examples{
plot(1:1000,1:1000, col=white.black(1000), pch=19)
}
\seealso{
\link[grDevices]{colorRampPalette}
}

