% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xSec.R
\name{xSec}
\alias{xSec}
\title{xSec}
\usage{
xSec(ve, vn, vu, transectHeading)
}
\arguments{
\item{ve}{Vector of east velocities within cells}

\item{vn}{Vector of north velocities within cells. Must be the same length as \code{ve}.}

\item{vu}{Vector of vertical velocities within cells. Must be the same length as \code{ve}.}

\item{transectHeading}{Compass heading of transect, assumes transect begins at river right and heads towards river left.}
}
\value{
Returns a \link[base]{data.frame} with compass heading of flow perpendicular to cross-section (vxHeading), compass heading of flow parallel to cross-section (vyCompassHeading), arithmetic heading of flow parallel to cross-section (vyArithmeticHeading), difference between flow perpendicular to cross-section and heading of flow within a cell (diffHeading), velocity perpendicular to cross-section (vx), and velocity parallel to cross-section where positive values indicate flow from river right to river left (vy).
}
\description{
Rotates velocities to be parallel and perpendicular to a cross-section
}
\examples{
data(cellVels)
velSub <- cellVels[transectName == "t1",,]
tHeading <- transectHeading(velSub$UTM_X_Proj, velSub$UTM_Y_Proj, velSub$Vel.E, velSub$Vel.N, velSub$cellDepth)
secondary <- xSec(velSub$Vel.E, velSub$Vel.N, velSub$Vel.up, tHeading)
}
\seealso{
\link{transectHeading}
}

