% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zeroSecQ.R
\name{zeroSecQ}
\alias{zeroSecQ}
\title{zeroSecQ}
\usage{
zeroSecQ(ve, vn, vu, cellHeight, cellWidth, transectHeading)
}
\arguments{
\item{ve}{Vector of east velocities within cells}

\item{vn}{Vector of north velocities within cells. Must be the same length as \code{ve}.}

\item{vu}{Vector of vertical velocities within cells. Must be the same length as \code{ve}.}

\item{cellHeight}{Vector of cell heights. Must be the same length as \code{ve}.}

\item{cellWidth}{Vector of cell Widths. Must be the same length as \code{ve}.}

\item{transectHeading}{Compass heading of transect, assumes transect begins at river right and heads towards river left.}
}
\value{
Returns a \link[base]{data.frame} with within cell discharge perpendicular to cross-section (qx), within cell discharge parallel to cross-section (qy), within cell discharge in the primary flow direction (qp), within cell discharge in the secondary flow direction (qs), velocity in the primary direction (vp.zsq), velocity in the secondary direction (vs.zsq), compass heading of primary velocity (vpHeading.zsq), within cell compass heading of secondary velocity (vsHeading.zsq), and arithmetic heading of within cell secondary velocity (vsArithmeticHeading.zsq).
}
\description{
Uses the zero secondary discharge method of rotation to calculate secondary velocities within a cross-section
}
\examples{
data(cellVels)
velSub <- cellVels[transectName == "t1",,]
tHeading <- transectHeading(velSub$UTM_X_Proj, velSub$UTM_Y_Proj, velSub$Vel.E, velSub$Vel.N, velSub$cellDepth)
zsq <- zeroSecQ(velSub$Vel.E, velSub$Vel.N, velSub$Vel.up, velSub$cellHeight, velSub$cellWidth, tHeading)
}
\references{
Lane, S.N., K.F. Bradbrook, K.S. Richards, P.M. Biron, and A.G. Roy. 2000. Secondary circulation cells in river channel confluences: Measurement artefacts or coherent flow structures? Hydrological Processes 14:2047-2071.
}
\seealso{
\link{xSec}.
}

