% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heading.R
\name{heading}
\alias{heading}
\title{heading}
\usage{
heading(ve, vn)
}
\arguments{
\item{ve}{A vector of velocities to the east.}

\item{vn}{A vector of velocities to the north. Must be the same length as ve}
}
\value{
A vector the same length as \code{ve}.
}
\description{
Calculates the degree heading of two velocity vectors
}
\examples{
heading(rnorm(5), rnorm(5))
easting <- c(0,1,1,1,0,-1,-1,-1)
northing <- c(1,1,0,-1,-1,-1,0,1)
heading(easting,northing)
}

